# docs/swagger.yaml
basePath: /
definitions:
...
  internal_delivery_http.AuthResponse:
    properties:
      message:
        example: User registered successfully
        type: string
    type: object
  internal_delivery_http.ErrorResponse:
    properties:
      error:
        example: invalid email format
        type: string
    type: object
  internal_delivery_http.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  internal_delivery_http.LoginResponse:
    properties:
      token:
        example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_delivery_http.ProfileResponse:
    properties:
      created_at:
        example: "2026-03-23T10:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      provider:
        example: local
        type: string
      updated_at:
        example: "2026-03-23T10:00:00Z"
        type: string
    type: object
  internal_delivery_http.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Skiix authentication API with JWT and OAuth2 support
  title: Skiix API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            $ref: '#/definitions/internal_delivery_http.LoginResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/internal_delivery_http.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - auth
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Retrieve the current authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            $ref: '#/definitions/internal_delivery_http.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Register credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/internal_delivery_http.AuthResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
    x-tokenname: Authorization
swagger: "2.0"